pyproject.toml
[project]
name            = "mcp-amadeus"
version         = "0.1.0"
description     = "Add your description here"
readme          = "README.md"
requires-python = ">=3.13"

dependencies = [
    "amadeus>=12.0.0",   # hotel endpoints
    "mcp[cli]>=1.6.0",   # FastMCP runtime/CLI
    "requests>=2.31",    # raw HTTP calls for flight offers
    "python-dotenv>=1.0" # load .env with API keys
]
